% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{stern_brocot_tree}
\alias{stern_brocot_tree}
\title{stern_brocot_tree}
\usage{
stern_brocot_tree(x)
}
\arguments{
\item{x}{The depth of the tree to generate.}
}
\value{
A data frame containing:
  - `num`: Numerator of the fraction.
  - `den`: Denominator of the fraction.
  - `depth`: The depth level in the Stern-Brocot tree.
}
\description{
Generate the Stern-Brocot tree up to a given depth using `std::list`.
}
